# shared/openapi/openapi.yaml
openapi: 3.0.2
paths:
...
  /workers/dxnet/jobs/next:
    post:
      tags: *a5
      parameters: []
      operationId: job.next
      requestBody:
        description: Body
        content:
          application/json:
            schema:
              type: object
              properties:
                botUserFriendCode:
                  type: string
                  minLength: 1
                waitMs:
                  type: integer
                  minimum: 0
                  maximum: 25000
              required:
                - botUserFriendCode
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  friendCode:
                    type: string
                  jobType:
                    type: string
                    enum:
                      - send_friend_request
                      - accept_friend_request
                      - update_score
                      - get_user_recent_event
                  priority:
                    type: integer
                  skipUpdateScore:
                    type: boolean
                  botUserFriendCode:
                    type: string
                    nullable: true
                  friendRequestSentAt:
                    type: string
                    nullable: true
                  friendRequestWaitStartedAt:
                    type: string
                    nullable: true
                  status:
                    type: string
                    enum:
                      - queued
                      - processing
                      - completed
                      - failed
                      - canceled
                  stage:
                    type: string
                    enum:
                      - send_request
                      - wait_acceptance
                      - wait_user_request
                      - accept_request
                      - update_score
                      - get_user_recent_event
                  profile:
                    nullable: true
                  error:
                    type: string
                    nullable: true
                  executing:
                    type: boolean
                  scoreProgress:
                    type: object
                    properties:
                      completedDiffs:
                        type: array
                        items:
                          type: integer
                          minimum: 0
                          maximum: 14
                      totalDiffs:
                        type: integer
                        minimum: 0
                    required:
                      - completedDiffs
                      - totalDiffs
                    nullable: true
                  updateScoreDuration:
                    type: number
                    nullable: true
                  autoExportResult:
                    type: object
                    properties:
                      divingFish:
                        type: object
                        properties:
                          status:
                            type: string
                          message:
                            type: string
                        required:
                          - status
                        nullable: true
                      lxns:
                        type: object
                        properties:
                          status:
                            type: string
                          message:
                            type: string
                        required:
                          - status
                        nullable: true
                    nullable: true
                  isAuthenticated:
                    type: boolean
                  cabinetScoreMap:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        achievement:
                          type: number
                        dxScore:
                          type: number
                      required:
                        - achievement
                        - dxScore
                    nullable: true
                  diffsToScrape:
                    type: array
                    items:
                      type: integer
                    nullable: true
                  runAt:
                    type: string
                    nullable: true
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
                required:
                  - id
                  - friendCode
                  - jobType
                  - skipUpdateScore
                  - status
                  - stage
                  - createdAt
                  - updatedAt
        "204":
          description: "204"
          content:
            application/json:
              schema: {}
  /workers/dxnet/jobs/{jobId}:
    patch:
      tags: *a5
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      operationId: job.patch
      requestBody:
        description: Body
        content:
          application/json:
            schema:
              type: object
              properties:
                botUserFriendCode:
                  type: string
                  nullable: true
                status:
                  type: string
                  enum:
                    - queued
                    - processing
                    - completed
                    - failed
                    - canceled
                stage:
                  type: string
                  enum:
                    - send_request
                    - wait_acceptance
                    - wait_user_request
                    - accept_request
                    - update_score
                    - get_user_recent_event
                result:
                  nullable: true
                profile:
                  nullable: true
                error:
                  type: string
                  nullable: true
                friendRequestSentAt:
                  type: string
                  nullable: true
                friendRequestWaitStartedAt:
                  type: string
                  nullable: true
                executing:
                  type: boolean
                runAt:
                  type: string
                  nullable: true
                updatedAt:
                  type: string
                scoreProgress:
                  type: object
                  properties:
                    completedDiffs:
                      type: array
                      items:
                        type: integer
                        minimum: 0
                        maximum: 14
                    totalDiffs:
                      type: integer
                      minimum: 0
                  required:
                    - completedDiffs
                    - totalDiffs
                  nullable: true
                addCompletedDiff:
                  type: integer
                  minimum: 0
                updateScoreDuration:
                  type: number
                  nullable: true
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  friendCode:
                    type: string
                  jobType:
                    type: string
                    enum:
                      - send_friend_request
                      - accept_friend_request
                      - update_score
                      - get_user_recent_event
                  priority:
                    type: integer
                  skipUpdateScore:
                    type: boolean
                  botUserFriendCode:
                    type: string
                    nullable: true
                  friendRequestSentAt:
                    type: string
                    nullable: true
                  friendRequestWaitStartedAt:
                    type: string
                    nullable: true
                  status:
                    type: string
                    enum:
                      - queued
                      - processing
                      - completed
                      - failed
                      - canceled
                  stage:
                    type: string
                    enum:
                      - send_request
                      - wait_acceptance
                      - wait_user_request
                      - accept_request
                      - update_score
                      - get_user_recent_event
                  profile:
                    nullable: true
                  error:
                    type: string
                    nullable: true
                  executing:
                    type: boolean
                  scoreProgress:
                    type: object
                    properties:
                      completedDiffs:
                        type: array
                        items:
                          type: integer
                          minimum: 0
                          maximum: 14
                      totalDiffs:
                        type: integer
                        minimum: 0
                    required:
                      - completedDiffs
                      - totalDiffs
                    nullable: true
                  updateScoreDuration:
                    type: number
                    nullable: true
                  autoExportResult:
                    type: object
                    properties:
                      divingFish:
                        type: object
                        properties:
                          status:
                            type: string
                          message:
                            type: string
                        required:
                          - status
                        nullable: true
                      lxns:
                        type: object
                        properties:
                          status:
                            type: string
                          message:
                            type: string
                        required:
                          - status
                        nullable: true
                    nullable: true
                  isAuthenticated:
                    type: boolean
                  cabinetScoreMap:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        achievement:
                          type: number
                        dxScore:
                          type: number
                      required:
                        - achievement
                        - dxScore
                    nullable: true
                  diffsToScrape:
                    type: array
                    items:
                      type: integer
                    nullable: true
                  runAt:
                    type: string
                    nullable: true
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
                required:
                  - id
                  - friendCode
                  - jobType
                  - skipUpdateScore
                  - status
                  - stage
                  - createdAt
                  - updatedAt
...
  /workers/sdgb/jobs/next:
    post:
      tags: &a10
        - sdgbWorker
      parameters: []
      operationId: sdgbWorker.next
      requestBody:
        description: Body
        content:
          application/json:
            schema:
              type: object
              properties:
                workerId:
                  type: string
                  minLength: 1
              required:
                - workerId
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  jobType:
                    type: string
                    enum:
                      - scan_qr
                      - get_rival_hash
                      - add_rival
                  status:
                    type: string
                    enum:
                      - queued
                      - processing
                      - completed
                      - failed
                  payload:
                    type: object
                    additionalProperties: {}
                  result:
                    type: object
                    additionalProperties: {}
                    nullable: true
                  error:
                    type: string
                    nullable: true
                  requesterTag:
                    type: string
                    nullable: true
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
                required:
                  - id
                  - jobType
                  - status
                  - payload
                  - result
                  - error
                  - createdAt
                  - updatedAt
        "204":
          description: "204"
  /workers/sdgb/jobs/{jobId}:
    get:
      tags: *a10
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      operationId: sdgbWorker.get
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  jobType:
                    type: string
                    enum:
                      - scan_qr
                      - get_rival_hash
                      - add_rival
                  status:
                    type: string
                    enum:
                      - queued
                      - processing
                      - completed
                      - failed
                  payload:
                    type: object
                    additionalProperties: {}
                  result:
                    type: object
                    additionalProperties: {}
                    nullable: true
                  error:
                    type: string
                    nullable: true
                  requesterTag:
                    type: string
                    nullable: true
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
                required:
                  - id
                  - jobType
                  - status
                  - payload
                  - result
                  - error
                  - createdAt
                  - updatedAt
        "404":
          description: "404"
    patch:
      tags: *a10
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      operationId: sdgbWorker.patch
      requestBody:
        description: Body
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    - queued
                    - processing
                    - completed
                    - failed
                result:
                  type: object
                  additionalProperties: {}
                  nullable: true
                error:
                  type: string
                  nullable: true
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  jobType:
                    type: string
                    enum:
                      - scan_qr
                      - get_rival_hash
                      - add_rival
                  status:
                    type: string
                    enum:
                      - queued
                      - processing
                      - completed
                      - failed
                  payload:
                    type: object
                    additionalProperties: {}
                  result:
                    type: object
                    additionalProperties: {}
                    nullable: true
                  error:
                    type: string
                    nullable: true
                  requesterTag:
                    type: string
                    nullable: true
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
                required:
                  - id
                  - jobType
                  - status
                  - payload
                  - result
                  - error
                  - createdAt
                  - updatedAt
        "404":
          description: "404"
info:
  title: Maimai Score Hub API
  version: 1.0.0
  description: Generated from @maimai-score-hub/shared ts-rest contracts
servers:
  - url: /api/v1